Use a protocol extension to handle the error
