+++ Lock/MultifactorAuthenticatable.swift
@@ -22,7 +22,7 @@
 
 import Foundation
 
-protocol MultifactorAuthenticatable {
+protocol MultifactorAuthenticatable: CredentialAuthenticatable {
     var code: String? { get }
 
     mutating func setMultifactorCode(_ code: String?) throws
